scripts: builddeb: replace newly built dtb only
It was removing every dtbs in the folder so while configuring the
extlinux it was replacing dtb and trying to change the extlinux
fdt variable.

Instead of that it will now look for the given dtb only and remove that
before installing and copy after the new one after installation.

There will be link of the rk-kernel.dtb in the rootfs by default so
it will just take this dtb from rootfs only.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>