scripts: builddeb: mount boot partition to /boot
Making mountboot.sh which mounts boot partition to /boot by finding
current root device and gets boot partition of that device from
the `PARTLABEL`.

Before this, there was already mountboot.h in /etc/init.d but that
is what is supplied with the rootfs, that rootfs script has problem
in identifying boot partition while there are multiple boot partitions
are there.

It was mounting other device's boot partition to /boot from which
the system has not booted.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>