arm64: dts: vaaman: include fpga communication overlay into rk3399-vaaman dtsi
The rk3399-vaaman-fpga-communication overlay was enabled by default
in the image, there is no meaning of turning that overlay off
as for continuous programming FPGA this is necessary.

Moving the contents of it into the rk3399-vaaman.dtsi.

So now, that overlay has been removed and the things which it carried
will be enabled by default.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>